move buildDLCWalletCallbacks to WebsocketUtil
